More sensor data
Now pulling in magnetic and environmental data in addition to the LED state of our own board. Next up is figuring out if I can access clean data from the underlying chips or if I have to just accept the data I get from the FreeSixIMU library.

Figured out the issue with the serial dropouts was due to the arduino running as fast as it could. I added a simple 100 microsecond delay to loop() and now serial data is transferred flawlessly.