Magnetic calibration and serial stability improvements
Starting with the stability improvements, I had noticed when looking at the serial monitor that some of the data being over was split over multiple lines. Its as if sometimes the data pulled from the sensor actually contains its own \r\n. I decided to just add a unique identifier that I knew could never end up in a bytestream and I use this value to be absolutely sure the message is complete. It seems to be working better so far.

Magnetic calibration is sort of ok. The actual calibration with respect to the Earth's magnetic field is great. Its zerod and scaled together. There is still an asymmetry however with respect to the north and south poles on each axis. The most obvious is the Z axis which had a massive bias. One pole of the magnet would be detected at 6 inches while the other pole had to be at 3 inches. We can compensate for this when we know the board is flat but not when its tilted in any way. We do this basic compensation when in polarity mode but it will only be accurate if its in the same position it was during the resting portion of calibration (red and blue lights).

This is a step towards adding tilt compensation to the magnetic heading data. Now that the data from each axis is normalized and aligned it can be used with the gyro/accel data to correct for tilt.