Cleaned up the code
Cleaned up a lot of the arduino side of the code by separating the different sensor data into different structs with their own functions to get that data. The final struct we send has the other structs inside it. Looks a bit nicer now and will be easier to go back and figure out how this all works in the future :P . It also doesn't double calculate the RGB stuff anymore. Finally, I added a couple bools to control whether the magnetic calibration is used and whether we try and set a dynamic ambient min.