First try at tilt compensation
Holy smokes trig always beats the hell out of me. I've got it mostly working but theres a small error somewhere in the maths representing the rotation matrix im using. The matrix is a combination of both the X and Y rotations so I could probably just separate those operations to isolate the issue. Interestingly, if you swap the order in the arctan function from y,x to x,y you swap the error from the roll to the pitch axis. I'm sure thats a dead giveaway as to where the problem is but my brain is not having it.

However, even these results are still impressive. The error is reduced by a huge amount to less than ~10-15 degrees at the worst (where before it could be off by 50+ degrees). An important note is the order of operation for the correcting factors. You must realign the reference frames BEFORE applying calibration factors, otherwise things don't go well.

The issue I'm having now is there is a ton of noise. The heading can jump around quite a lot so some smoothing will be needed. Ideally I would have an output that tracked the average value over the last N seconds and use that. Maybe PID if we want to get really fancy although thats probably overkill.