Implemented magnetic heading smoothing
Got the heading filter working. At first I tried saving the raw heading value and took an average over the last N values to get a smoothed heading. This worked fantastically well except around north which can be a 0 or a 360. Swapping between 0 and 360 rapidly just completely messed up the averaging so I had to use a periodic function. After a lot of testing I settled on a sin/cos/atan2 combo. I tried to just use sin or cos with their inverse functions but the output was unusable. My trig isnt great so I just kept trying function combos until I got one that worked. The next solution to try is to define an error function and implement PID control or parts of it.

Also moved the magnetic hard/soft iron deposit compensation to be applied just before we calculate our magnetic heading. Previously we were only applying the calibration after getting the heading so it only applied to the lightshow and GUI data.

The app is pretty much complete now I think. Maybe it would be nice to be able to control the mode from the GUI. Possibly add more widgets to the GUI like a compass or add graphing. Actually graphing is a great idea, maybe that will be next.