Fix bug and change to calibration
Some of the code was copied from a different project im working on and a few issues cropped up because of that. First off mag2 doesnt exist so that was a problem. Secondly, it was applying the tilt compensation to the raw magnetic field data which we don't want. We just want the heading tilt compensated.